backend/requirements.txt
fastapi
uvicorn
pydantic
torch
opencv-python
numpy
pillow
